LPD-29276 Extract handleMultiSelectRoleItemsChange function to a util file and reuse it to avoid code repetition
