+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/components/SettingsContainer/PrimaryRecipients.tsx
@@ -21,7 +21,7 @@ import {
 import React, {useEffect, useState} from 'react';
 
 import {NotificationTemplateError} from '../EditNotificationTemplate';
-import {getCheckedChildren} from './rolesUtils';
+import {getCheckedChildren, handleMultiSelectRoleItemsChange} from './rolesUtils';
 
 interface PrimaryRecipientProps {
 	emailNotificationRoles: MultiSelectItem[];
@@ -44,30 +44,6 @@ export function PrimaryRecipient({
 }: PrimaryRecipientProps) {
 	const [recipient] = values.recipients as EmailRecipients[];
 	const [toRolesList, setToRolesList] = useState<MultiSelectItem[]>([]);
-
-	const handleMultiSelectItemsChange = (itemsGroup: MultiSelectItem[]) => {
-		const newRecipients: EmailNotificationRecipients[] = [];
-
-		if (itemsGroup.length) {
-			itemsGroup.forEach((itemGroup) => {
-				itemGroup.children.forEach((child) => {
-					if (child.checked) {
-						newRecipients.push({['roleName']: child.value});
-					}
-				});
-			});
-		}
-
-		setValues({
-			...values,
-			recipients: [
-				{
-					...recipient,
-					to: newRecipients,
-				},
-			],
-		});
-	};
 
 	useEffect(() => {
 		if (emailNotificationRoles.length && !toRolesList.length) {
@@ -171,7 +147,18 @@ export function PrimaryRecipient({
 						)}
 						selectAllOption
 						setOptions={(items) => {
-							handleMultiSelectItemsChange(items);
+							const newRecipients = handleMultiSelectRoleItemsChange(items);
+
+							setValues({
+								...values,
+								recipients: [
+									{
+										...recipient,
+										to: newRecipients,
+									},
+								],
+							});
+						
 							setToRolesList(items);
 						}}
 					/>
